# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Snipo API
  description: |
    REST API for Snipo - a lightweight, self-hosted snippet manager.
    
    ## Authentication
    
    All protected endpoints require authentication via one of:
    - **Session Cookie**: `snipo_session` cookie (for web UI)
    - **Bearer Token**: `Authorization: Bearer <token>` header
    - **API Key**: `X-API-Key: <key>` header
    
    API tokens can be created via the Settings page or `/api/v1/tokens` endpoint.
    
    ### Token Permissions
    
    API tokens have permission levels that control access:
    - **read**: Read snippets, tags, folders, history, and Gist status
    - **write**: Read access plus create, update, delete, restore, and sync operations
    - **admin**: Write access plus token management, settings, backups, and Gist configuration
    
    Session-based auth (web UI) has full admin access by default.
    
    ## Rate Limiting
    
    API endpoints are rate-limited based on permission level:
    - **Read operations**: 1000 requests/hour
    - **Write operations**: 500 requests/hour
    - **Admin operations**: 100 requests/hour
    
    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Requests remaining in current window
    - `X-RateLimit-Reset`: Unix timestamp when the limit resets
    - `Retry-After`: Seconds to wait before retrying (when limit exceeded)
    
...
  version: 1.0.0
  license:
...
  contact:
...
servers:
...
paths:
...
  /api/v1/gist/sync/enable-all:
    post:
      tags: [GitHub Gist Sync]
      summary: Enable sync for all snippets
      description: Creates GitHub gists for all snippets and enables sync
      operationId: enableSyncForAll
      security:
        - sessionCookie: []
        - bearerAuth: []
        - apiKey: []
      responses:
        '200':
          description: Sync enabled for snippets
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  enabled:
                    type: integer
                  errors:
                    type: integer
                  error_messages:
                    type: array
                    items:
                      type: string
        '400':
          description: Sync not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                not_configured:
                  summary: GitHub sync not configured
                  value:
                    error:
                      code: "SYNC_NOT_CONFIGURED"
                      message: "GitHub token not configured"
                fetch_failed:
                  summary: Failed to fetch snippets
                  value:
                    error:
                      code: "FETCH_FAILED"
                      message: "Failed to fetch snippets"
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                no_auth:
                  summary: No authentication provided
                  value:
                    error:
                      code: "UNAUTHORIZED"
                      message: "Authentication required"
                invalid_token:
                  summary: Invalid or expired token
                  value:
                    error:
                      code: "INVALID_TOKEN"
                      message: "Invalid or expired authentication token"
        '403':
          description: Forbidden - write permission required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                insufficient_permissions:
                  summary: Write permission required
                  value:
                    error:
                      code: "FORBIDDEN"
                      message: "Insufficient permissions to perform this action"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                database_error:
                  summary: Failed to enable sync
                  value:
                    error:
                      code: "INTERNAL_ERROR"
                      message: "An internal server error occurred"
...
  /api/v1/gist/sync/enable/{id}:
    post:
      tags: [GitHub Gist Sync]
      summary: Enable sync for a snippet
      description: Creates a GitHub gist for the snippet and enables sync
      operationId: enableSyncForSnippet
      security:
        - sessionCookie: []
        - bearerAuth: []
        - apiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Snippet ID
      responses:
        '200':
          description: Sync enabled
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                sync_not_configured:
                  summary: Sync not configured
                  value:
                    error:
                      code: "SYNC_NOT_CONFIGURED"
                      message: "GitHub token not configured"
                snippet_not_found:
                  summary: Snippet not found
                  value:
                    error:
                      code: "NOT_FOUND"
                      message: "Snippet not found"
                missing_id:
                  summary: Missing snippet ID
                  value:
                    error:
                      code: "MISSING_ID"
                      message: "Snippet ID is required"
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                no_auth:
                  summary: No authentication provided
                  value:
                    error:
                      code: "UNAUTHORIZED"
                      message: "Authentication required"
                invalid_token:
                  summary: Invalid or expired token
                  value:
                    error:
                      code: "INVALID_TOKEN"
                      message: "Invalid or expired authentication token"
        '403':
          description: Forbidden - write permission required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                insufficient_permissions:
                  summary: Write permission required
                  value:
                    error:
                      code: "FORBIDDEN"
                      message: "Insufficient permissions to perform this action"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                enable_failed:
                  summary: Failed to enable sync
                  value:
                    error:
                      code: "ENABLE_FAILED"
                      message: "Failed to enable sync for snippet"
  /api/v1/gist/sync/disable/{id}:
    post:
      tags: [GitHub Gist Sync]
      summary: Disable sync for a snippet
      description: Disables sync for the snippet (does not delete the gist)
      operationId: disableSyncForSnippet
      security:
        - sessionCookie: []
        - bearerAuth: []
        - apiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Snippet ID
      responses:
        '200':
          description: Sync disabled
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                disable_failed:
                  summary: Failed to disable sync
                  value:
                    error:
                      code: "DISABLE_FAILED"
                      message: "Failed to disable sync for snippet"
                sync_not_configured:
                  summary: Sync not configured
                  value:
                    error:
                      code: "SYNC_NOT_CONFIGURED"
                      message: "GitHub token not configured"
                missing_id:
                  summary: Missing snippet ID
                  value:
                    error:
                      code: "MISSING_ID"
                      message: "Snippet ID is required"
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                no_auth:
                  summary: No authentication provided
                  value:
                    error:
                      code: "UNAUTHORIZED"
                      message: "Authentication required"
                invalid_token:
                  summary: Invalid or expired token
                  value:
                    error:
                      code: "INVALID_TOKEN"
                      message: "Invalid or expired authentication token"
        '403':
          description: Forbidden - write permission required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                insufficient_permissions:
                  summary: Write permission required
                  value:
                    error:
                      code: "FORBIDDEN"
                      message: "Insufficient permissions to perform this action"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                disable_failed:
                  summary: Failed to disable sync
                  value:
                    error:
                      code: "DISABLE_FAILED"
                      message: "Failed to disable sync for snippet"
  /api/v1/gist/sync/snippet/{id}:
    post:
      tags: [GitHub Gist Sync]
      summary: Sync a specific snippet
      description: Syncs a specific snippet with its GitHub gist
      operationId: syncSnippet
      security:
        - sessionCookie: []
        - bearerAuth: []
        - apiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Snippet ID
      responses:
        '200':
          description: Snippet synced
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                sync_not_configured:
                  summary: Sync not configured
                  value:
                    error:
                      code: "SYNC_NOT_CONFIGURED"
                      message: "GitHub token not configured"
                snippet_not_found:
                  summary: Snippet not found
                  value:
                    error:
                      code: "NOT_FOUND"
                      message: "Snippet not found"
                missing_id:
                  summary: Missing snippet ID
                  value:
                    error:
                      code: "MISSING_ID"
                      message: "Snippet ID is required"
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                no_auth:
                  summary: No authentication provided
                  value:
                    error:
                      code: "UNAUTHORIZED"
                      message: "Authentication required"
                invalid_token:
                  summary: Invalid or expired token
                  value:
                    error:
                      code: "INVALID_TOKEN"
                      message: "Invalid or expired authentication token"
        '403':
          description: Forbidden - write permission required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                insufficient_permissions:
                  summary: Write permission required
                  value:
                    error:
                      code: "FORBIDDEN"
                      message: "Insufficient permissions to perform this action"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                sync_failed:
                  summary: Failed to sync snippet
                  value:
                    error:
                      code: "SYNC_FAILED"
                      message: "Failed to sync snippet"

    delete:
      tags: [GitHub Gist Sync]
      summary: Delete snippet gist
      description: Deletes the corresponding GitHub gist for a snippet and removes the sync mapping
      operationId: deleteSnippetGist
      security:
        - sessionCookie: []
        - bearerAuth: []
        - apiKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Snippet ID
      responses:
        '204':
          description: Gist deleted and mapping removed
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missing_id:
                  summary: Missing snippet ID
                  value:
                    error:
                      code: "MISSING_ID"
                      message: "Snippet ID is required"
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - write permission required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                delete_failed:
                  summary: Failed to delete gist
                  value:
                    error:
                      code: "DELETE_FAILED"
                      message: "Failed to delete gist"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: snipo_session
      description: |
        Session-based authentication for web UI.
        Full admin access (equivalent to admin permission level).
    bearerAuth:
      type: http
      scheme: bearer
      description: API token sent as an HTTP Bearer credential.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API token sent through the X-API-Key header.
  schemas:
    Error:
      type: object
      description: |
        Standard error response format. All errors include a code and message.
        
        **Common Error Codes:**
        
        **Authentication Errors (401):**
        - `UNAUTHORIZED`: Missing or invalid authentication credentials
        - `INVALID_TOKEN`: API token is invalid or expired
        - `SESSION_EXPIRED`: Session has expired, login required
        
        **Authorization Errors (403):**
        - `FORBIDDEN`: Insufficient permissions for this operation
        - `READ_ONLY_TOKEN`: Token has read-only permissions, write access required
        - `ADMIN_REQUIRED`: Operation requires admin-level permissions
        
        **Resource Errors (404):**
        - `NOT_FOUND`: Requested resource does not exist
        - `SNIPPET_NOT_FOUND`: Snippet with specified ID not found
        - `FOLDER_NOT_FOUND`: Folder with specified ID not found
        - `TAG_NOT_FOUND`: Tag with specified ID not found
        - `TOKEN_NOT_FOUND`: API token with specified ID not found
        - `HISTORY_NOT_FOUND`: History version not found
        
        **Rate Limiting (429):**
        - `RATE_LIMIT_EXCEEDED`: Too many requests, please slow down
          - Response includes `Retry-After` header with seconds to wait
          - Default limits: 100 req/min (authenticated), 20 req/min (public)
        
        **Server Errors (500):**
        - `INTERNAL_ERROR`: Unexpected server error occurred
        - `DATABASE_ERROR`: Database operation failed
        - `S3_ERROR`: S3 storage operation failed
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code (see description for full list)
              examples:
                - "UNAUTHORIZED"
            message:
              type: string
              description: Human-readable error message
              examples:
                - "Authentication required"
      examples:
        - error:
            code: "FORBIDDEN"
            message: "Insufficient permissions to perform this action"
...